requirements.txt
appdirs==1.4.4
argcomplete==3.5.0
importlib_resources==6.4.4
pillow==10.4.0
pypng==0.20220715.0
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.2.1
PyYAML==6.0.2
qrcode==7.4.2
six==1.16.0
typing_extensions==4.12.2
